requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0

# Templates and Static Files
jinja2>=3.1.0
aiofiles>=23.0.0

# HTTP Requests (for web scraping in lookup feature)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Form handling
python-multipart>=0.0.6

# Image processing (for favicon generation)
pillow>=10.0.0

# Type hints (Python < 3.9 compatibility)
typing-extensions>=4.8.0